requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
djangorestframework==3.6.3
httpie==0.9.9
idna==2.5
Pygments==2.2.0
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
